support/benchmark/pyproject.toml
[project]
name = "benchmark"
version = "0.1.0"
description = "MygramDB vs MySQL FULLTEXT benchmark tool"
authors = [
    { name = "libraz", email = "libraz@libraz.net" }
]
dependencies = [
    "mysql-connector-python>=9.5.0",
]
readme = "README.md"
requires-python = ">= 3.10"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "flake8>=7.3.0",
    "mypy>=1.18.2",
]
